requirements.txt
Django==4.2.7
dj-rest-auth==4.0.1
django-allauth==0.50.0
django-cors-headers==4.0.0
djangorestframework_simplejwt==5.2.2